Move config into _utils
