+++ src/dapla_metadata/variable_definitions/variable_definition.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 from pydantic import PrivateAttr
 
-from dapla_metadata.variable_definitions import config
+from dapla_metadata.variable_definitions._utils import config
 from dapla_metadata.variable_definitions._utils._client import VardefClient
 from dapla_metadata.variable_definitions._utils.variable_definition_files import (
     create_variable_yaml,
