+++ tests/variable_definitions/test_config.py
@@ -4,15 +4,21 @@ from dapla_metadata._shared.config import DAPLA_ENVIRONMENT
 from dapla_metadata._shared.config import DAPLA_GROUP_CONTEXT
 from dapla_metadata._shared.config import OIDC_TOKEN
 from dapla_metadata._shared.enums import DaplaEnvironment
-from dapla_metadata.variable_definitions.config import VARDEF_DEFAULT_DESCRIPTION_PATH
-from dapla_metadata.variable_definitions.config import VARDEF_DESCRIPTIONS_FILE_PATH
-from dapla_metadata.variable_definitions.config import VARDEF_HOST_TEST
-from dapla_metadata.variable_definitions.config import WORKSPACE_DIR
-from dapla_metadata.variable_definitions.config import get_active_group
-from dapla_metadata.variable_definitions.config import get_descriptions_path
-from dapla_metadata.variable_definitions.config import get_vardef_client_configuration
-from dapla_metadata.variable_definitions.config import get_vardef_host
-from dapla_metadata.variable_definitions.config import get_workspace_dir
+from dapla_metadata.variable_definitions._utils.config import (
+    VARDEF_DEFAULT_DESCRIPTION_PATH,
+)
+from dapla_metadata.variable_definitions._utils.config import (
+    VARDEF_DESCRIPTIONS_FILE_PATH,
+)
+from dapla_metadata.variable_definitions._utils.config import VARDEF_HOST_TEST
+from dapla_metadata.variable_definitions._utils.config import WORKSPACE_DIR
+from dapla_metadata.variable_definitions._utils.config import get_active_group
+from dapla_metadata.variable_definitions._utils.config import get_descriptions_path
+from dapla_metadata.variable_definitions._utils.config import (
+    get_vardef_client_configuration,
+)
+from dapla_metadata.variable_definitions._utils.config import get_vardef_host
+from dapla_metadata.variable_definitions._utils.config import get_workspace_dir
 
 
 @pytest.mark.parametrize(
